fix(polls): address review findings — FU conditional data loss, a11y, SSR token, sync split
- Fix file_upload conditional questions silently deleted on save-after-reload:
  initializeFromApiData now re-attaches FU conditionals (top-level + sections)
  so the write path no longer DELETEs them on the next save (data loss).
- PollCard close-time labels now show locale-aware relative time
  (Intl.RelativeTimeFormat) instead of a raw datetime.
- Split poll-api-sync.ts (764 lines) into orchestrator + questions +
  conditionals + types modules, all under the 500-line limit, and replace
  pervasive `any` with typed Api*Shape interfaces + generated QuestionnaireSchema.
- PollUrlStrip aria-label is now translated (pollCard.shareUrlLabel).
- Replace native confirm() in PollCard delete with accessible ConfirmDialog.
- PollVoteForm "not authenticated" throw now shows an inline toast and is
  marked silent so the global handler doesn't double-toast.
- Poll edit page re-syncs scalar form fields from server after save.
- /api/auth/refresh no longer returns the refresh token in the JSON body.
- New-poll page uses the toISOString datetime helper for consistency.
