+++ messages/de.json
@@ -6111,7 +6111,7 @@
 		"votes_zero": "Noch keine Stimmen",
 		"votes_one": "1 Stimme",
 		"votes_other": "{count} Stimmen",
-		"closesIn": "Schließt in {duration}",
+		"closesIn": "Schließt {duration}",
 		"closedAgo": "Geschlossen {ago}",
 		"notOpenedYet": "Noch nicht geöffnet",
 		"openPoll": "Umfrage öffnen",
@@ -6120,6 +6120,7 @@
 		"copyLink": "Link kopieren",
 		"copySuccess": "Umfragelink in die Zwischenablage kopiert",
 		"copyError": "Fehler beim Kopieren des Links",
+		"shareUrlLabel": "Umfrage-Freigabelink",
 		"moreActions": "Weitere Aktionen",
 		"viewAsVoter": "Als Teilnehmer ansehen",
 		"close": "Umfrage schließen",
@@ -6232,6 +6233,7 @@
 		"submitVote": "Abstimmen",
 		"submitting": "Wird gesendet…",
 		"submitSuccess": "Deine Stimme wurde erfasst.",
+		"submitErrorAuth": "Deine Sitzung ist abgelaufen. Bitte melde dich erneut an, um abzustimmen.",
 		"submitErrorClosed": "Diese Umfrage wurde gerade geschlossen.",
 		"submitErrorIneligible": "Du darfst an dieser Umfrage nicht teilnehmen.",
 		"submitErrorConflict": "Du hast bereits an dieser Umfrage abgestimmt.",
