+++ messages/en.json
@@ -6111,7 +6111,7 @@
 		"votes_zero": "No votes yet",
 		"votes_one": "1 vote",
 		"votes_other": "{count} votes",
-		"closesIn": "Closes in {duration}",
+		"closesIn": "Closes {duration}",
 		"closedAgo": "Closed {ago}",
 		"notOpenedYet": "Not opened yet",
 		"openPoll": "Open poll",
@@ -6120,6 +6120,7 @@
 		"copyLink": "Copy link",
 		"copySuccess": "Poll link copied to clipboard",
 		"copyError": "Failed to copy link",
+		"shareUrlLabel": "Poll share URL",
 		"moreActions": "More actions",
 		"viewAsVoter": "View as voter",
 		"close": "Close poll",
@@ -6232,6 +6233,7 @@
 		"submitVote": "Submit vote",
 		"submitting": "Submitting…",
 		"submitSuccess": "Your vote was recorded.",
+		"submitErrorAuth": "Your session expired. Please sign in again to vote.",
 		"submitErrorClosed": "This poll just closed.",
 		"submitErrorIneligible": "You're not allowed to vote on this poll.",
 		"submitErrorConflict": "You already voted on this poll.",
