+++ messages/it.json
@@ -6111,7 +6111,7 @@
 		"votes_zero": "Ancora nessun voto",
 		"votes_one": "1 voto",
 		"votes_other": "{count} voti",
-		"closesIn": "Chiude tra {duration}",
+		"closesIn": "Chiude {duration}",
 		"closedAgo": "Chiuso {ago}",
 		"notOpenedYet": "Non ancora aperto",
 		"openPoll": "Apri sondaggio",
@@ -6120,6 +6120,7 @@
 		"copyLink": "Copia link",
 		"copySuccess": "Link del sondaggio copiato negli appunti",
 		"copyError": "Impossibile copiare il link",
+		"shareUrlLabel": "URL di condivisione del sondaggio",
 		"moreActions": "Altre azioni",
 		"viewAsVoter": "Visualizza come votante",
 		"close": "Chiudi sondaggio",
@@ -6232,6 +6233,7 @@
 		"submitVote": "Invia voto",
 		"submitting": "Invio in corso…",
 		"submitSuccess": "Il tuo voto è stato registrato.",
+		"submitErrorAuth": "La tua sessione è scaduta. Accedi di nuovo per votare.",
 		"submitErrorClosed": "Questo sondaggio è appena stato chiuso.",
 		"submitErrorIneligible": "Non sei autorizzato a votare in questo sondaggio.",
 		"submitErrorConflict": "Hai già votato in questo sondaggio.",
