+++ src/lib/components/polls/PollUrlStrip.svelte
@@ -24,7 +24,13 @@
 </script>
 
 <div class="flex flex-col gap-2 sm:flex-row">
-	<Input type="text" readonly value={url} aria-label="Poll share URL" class="font-mono text-xs" />
+	<Input
+		type="text"
+		readonly
+		value={url}
+		aria-label={m['pollCard.shareUrlLabel']()}
+		class="font-mono text-xs"
+	/>
 	<Button
 		type="button"
 		variant="default"
