+++ src/lib/utils/questionnaire-api-converters.ts
@@ -327,7 +327,7 @@ export function initializeFromApiData(questionnaire: any, rawQ: any): InitFromAp
 	}
 
 	// ===== Step 6: Attach conditional questions/sections to parent options =====
-	function attachConditionalQuestion(apiQ: any, type: 'mc' | 'ft') {
+	function attachConditionalQuestion(apiQ: any, type: 'mc' | 'ft' | 'fu') {
 		if (!apiQ.depends_on_option_id || !conditionalQuestionIds.has(apiQ.id)) return;
 
 		const parentOption = optionMap.get(apiQ.depends_on_option_id);
@@ -337,10 +337,14 @@ export function initializeFromApiData(questionnaire: any, rawQ: any): InitFromAp
 			);
 			if (existingIds.has(apiQ.id)) return;
 
-			const convertedQ =
-				type === 'mc'
-					? convertApiMcQuestion(apiQ, apiQ.order ?? 0)
-					: convertApiFtQuestion(apiQ, apiQ.order ?? 0);
+			let convertedQ: QuestionnaireQuestion;
+			if (type === 'mc') {
+				convertedQ = convertApiMcQuestion(apiQ, apiQ.order ?? 0);
+			} else if (type === 'ft') {
+				convertedQ = convertApiFtQuestion(apiQ, apiQ.order ?? 0);
+			} else {
+				convertedQ = convertApiFuQuestion(apiQ, apiQ.order ?? 0);
+			}
 			parentOption.conditionalQuestions = parentOption.conditionalQuestions || [];
 			parentOption.conditionalQuestions.push(convertedQ);
 
@@ -352,23 +356,20 @@ export function initializeFromApiData(questionnaire: any, rawQ: any): InitFromAp
 		}
 	}
 
-	// Conditional MC questions
-	for (const apiQ of q.multiplechoicequestion_questions || []) {
-		attachConditionalQuestion(apiQ, 'mc');
-	}
-	for (const apiSection of q.sections || []) {
-		for (const apiQ of apiSection.multiplechoicequestion_questions || []) {
-			attachConditionalQuestion(apiQ, 'mc');
+	// Conditional questions (MC / FT / FU), both top-level and inside sections.
+	const conditionalCollections: [string, 'mc' | 'ft' | 'fu'][] = [
+		['multiplechoicequestion_questions', 'mc'],
+		['freetextquestion_questions', 'ft'],
+		['fileuploadquestion_questions', 'fu']
+	];
+	for (const [collection, type] of conditionalCollections) {
+		for (const apiQ of q[collection] || []) {
+			attachConditionalQuestion(apiQ, type);
 		}
-	}
-
-	// Conditional FT questions
-	for (const apiQ of q.freetextquestion_questions || []) {
-		attachConditionalQuestion(apiQ, 'ft');
-	}
-	for (const apiSection of q.sections || []) {
-		for (const apiQ of apiSection.freetextquestion_questions || []) {
-			attachConditionalQuestion(apiQ, 'ft');
+		for (const apiSection of q.sections || []) {
+			for (const apiQ of apiSection[collection] || []) {
+				attachConditionalQuestion(apiQ, type);
+			}
 		}
 	}
 
@@ -390,7 +391,12 @@ export function initializeFromApiData(questionnaire: any, rawQ: any): InitFromAp
 				const sectionFt = (apiSection.freetextquestion_questions || [])
 					.filter((ftq: any) => !ftq.depends_on_option_id)
 					.map((ftq: any) => convertApiFtQuestion(ftq, ftq.order ?? 0));
-				convertedSection.questions = [...sectionMc, ...sectionFt].sort((a, b) => a.order - b.order);
+				const sectionFu = (apiSection.fileuploadquestion_questions || [])
+					.filter((fuq: any) => !fuq.depends_on_option_id)
+					.map((fuq: any) => convertApiFuQuestion(fuq, fuq.order ?? 0));
+				convertedSection.questions = [...sectionMc, ...sectionFt, ...sectionFu].sort(
+					(a, b) => a.order - b.order
+				);
 				parentOption.conditionalSections = parentOption.conditionalSections || [];
 				parentOption.conditionalSections.push(convertedSection);
 			}
